refactor(bezier): move private _bezier_*.py modules into bezier/ package
Converts bezier.py to bezier/__init__.py and moves the 5 private
_bezier_*.py implementation files into the package, mirroring the
bspline/ package structure.
