+++ src/pantr/_bspline_product_nd.py
@@ -19,7 +19,6 @@ from typing import TYPE_CHECKING
 import numpy as np
 import numpy.typing as npt
 
-from ._bezier_product import _bernstein_product_coefficients_nd
 from ._bspline_knots import _get_Bspline_num_basis_1D_impl, _get_unique_knots_and_multiplicity_impl
 from ._bspline_product import (
     _build_product_knot_vector,
@@ -29,6 +28,7 @@ from ._bspline_product import (
     _lookup_mults_in_space,
     _merge_interior_breakpoints,
 )
+from .bezier._bezier_product import _bernstein_product_coefficients_nd
 from .bspline._bspline_space_1d import BsplineSpace1D
 from .bspline._bspline_space_nd import BsplineSpace
